ゲームマネージャの作り直し修正
