+++ Assets/Scripts/GameManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a3e834b81c1617745a10b20437cd6165
