--- Assets/Scripts/JanChainGame.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1884faaf890a1f347aa1c4851a73c3fa
